Bugfix #21825
Auto-IK: applying back the IK transform on bones fails for constrainted
bones themselves. This would require the post-transform of IK to be
converted to a pre-transform for the constraint. That's too much of
a headache for bugfixing days.

To keep the autoIK feature work it now stops IK chains at constrainted
bones. Added todo item for a matrix wizzard to fix this one day.

(What would be really cool is a matrix-api for armature to easily
get all the required pre- and post transforms and apply. There's a lot
of duplicate exception handling for hinge bones for example.)