Fix/workaround for russian ui crash on exit
Crash was caused by data segment corruption near "string" in view3d_modeselect_pup().
Enlarged size of this static buffer, so it's now enough to store translated modeselect string.

It's not actually fixes because in some other language this modeline might be much longer
and it'll lead to corruptions again.