bge font objects: fix for difference in size when object has a parent for after 2.62 I would like to try if using doubles would help to increase the precision here.
Also now that the mipmap patch is reverted we may want to look at this problem from a different perspective.
RES could be calculated taking the object size in relation to the camera (i.e. size and distance, ignoring rotation).
That may solve both problems of wrong resolution and lack of smoothness.

For the time being users still need to use object.resolution to fine tune the rendered text.