Fix #33398: Missed undo push when script execution failed
Undo push was missed in cases when script failed to run with some error
when using Run Script in text editor. As far as i can see it makes sense
to skip undo push here only in cases live editing is enabled.

Otherwise it's indeed annoying to return to previous scene state when
debugging the script.